refactor(claude): extract stop delivery orchestration
